Moved CSS files to Style folder
