+++ frontend/src/Components/NavBar.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { FaBars, FaTimes } from 'react-icons/fa'
 import { Link } from "react-router-dom";
-import './NavBar.css'
+import '../Style/Navbar/NavBar.css'
 
 const Navbar = () => {
     const [click, setClick] = useState(false)
